All syscall banks can now be used as overlays; saves lots of space --- we have room for a 27kB userland now! Also moves unlinki and the guts of _sync into common code because they're called from multiple banks.
