+++ Kernel/tools/build.mk
@@ -1,0 +1,5 @@
+$(call find-makefile)
+
+map_syscall.srcs = map_syscall.c
+$(call build, map_syscall, host-exe)
+
